third_party/blink/web_tests/platform/generic/fast/dom/HTMLAnchorElement/set-href-attribute-rebase-expected.txt

Tests how setting href attributes affects setting a new base for the URL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS a.href is "http://old_base/foo?query"
PASS document.querySelector('a').href is "http://new_base/foo?query"
PASS successfullyParsed is true

TEST COMPLETE
